@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # Copyright (c) 2013-2018 Paulo Freitas
 # MIT License (see LICENSE file)
@@ -8,8 +8,6 @@ File formats package
 This package provides an abstract file format class, a file format factory
 class, a file format repository class and concrete file format modules.
 '''
-from __future__ import absolute_import
-
 # Imports
 
 # Built-in dependencies
